dao part is updated

+++ src/main/java/data/IUserFactory.java
@@ -1,11 +1,10 @@
 package data;
 
-import model.League;
 import model.User;
 
 public interface IUserFactory {
 
-    void loadUserByName(long id, User user) throws Exception;
+    void loadUserByName(int id, User user) throws Exception;
     long addUser(User user) throws Exception;
 
     }
